image_fetcher/requirements.txt
pymongo
torch
torchvision
requests
psycopg2-binary
requests
sqlalchemy
pandas